Do not subclass ImgurResponse from QMap
In the past, this class used two QMaps: one stored root-level values
from JSON response, and the other one stored sub-fields of "data" field.
The former QMap was in the inheritance chain, whereas the other
was aggregated as a class member.

This has changed recently, and map for root-level values is not in use
anymore.  In this commit, it is removed entirely.

Some leftovers (reading map which is always empty) are cleaned up too.